all: clean up and consolidate deployment documentation
This change moves the devapp deployment documentation into the doc
directory because it's common to all services that run in the GKE
cluster.

It also restructures that documentation just a little to alleviate some
confusion (i.e. make push-prod is not necessary if using make
deploy-prod).

Then, it updates all the relevant services' documentation to refer to
that shared document.

Also, it touch up a couple of the services' READMEs, like gopherbot and
pubsubhelper.

Change-Id: I3774385f1d0ba9e9886ece8462294c1f28e8bd17
Reviewed-on: https://go-review.googlesource.com/c/build/+/416220
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>