cmd/gopherbot: omit owner from assign reviewers
Assign reviewers correctly omits the owner from the list of gerrit
reviewers, but not from the list of metadata on the CL. When adding
a "Run-TryBot +1" vote to a CL, the owner is listed as a Reviewer in the
metadata.

This change updates both cases to use the same filter, simplifying the
humanReviewersOnChange to check for bots and owner the same way for
metadata and reviewers.

Fixes golang/go#53077

Change-Id: Ia3d4939f6fb8d81cfb371683025b5e3ce0282fd7
Reviewed-on: https://go-review.googlesource.com/c/build/+/417479
Run-TryBot: Jenny Rakoczy <jenny@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Jenny Rakoczy <jenny@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>