cmd/genbootstrap: update for regular use
A handful of changes to make this possible to use for updating
the toolchains on all builders.

- Add -rev flag to set revision of toolchain to build.
- Download specific revision from Gerrit instead of assuming $GOROOT is appropriate.
- Add revision suffix to tgz output names.
- Add -upload flag to upload to the right bucket.
- Add -verbose flag and make the output fairly quiet by default.
- Allow glob pattern to match known targets.

With these changes, in order to build and upload bootstrap toolchains
for Go 1.17.13 (the new default) for all toolchains:

	genbootstrap -upload '*'

Or to build a newer one for one system:

	genbootstrap -rev go1.19rc2 -upload linux-loong64

For golang/go#44505.

Change-Id: I09fe6408f91b15018e63af711e71af6576d5e6be
Reviewed-on: https://go-review.googlesource.com/c/build/+/419082
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Run-TryBot: Russ Cox <rsc@golang.org>