internal/workflow: make Dependencies less implicit and more flexible
It's difficult for workflow definition readers to tell whether an
argument to Task is a Value or a Dependency. In addition, it's
impossible to use a Value as just an ordering constraint. Currently the
only example of this friction is approval actions, which take an
extra argument only to ignore it.

Fix both problems with a new method, After, which consumes any number of
Values and Dependencies and produces a TaskInput that acts like a
Dependency used to. Adopt it everywhere.

Unfortunately, it's impossible to statically check this; Values have to
be usable as TaskInputs. I could introduce a parallel Arguments function
that bundles all the arguments together but that would add a lot of
clutter to the common case.

Change-Id: I276e09f4a90e6cf595280c086c481dd6f14de5ca
Reviewed-on: https://go-review.googlesource.com/c/build/+/419422
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Jenny Rakoczy <jenny@golang.org>
Auto-Submit: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>