cmd/greplogs: associate known issues by regex
Add a flag --known issue that associates issues with log entries based
on regexp matches. It's used like this:

--known-issue='#53456=TestDebugLines'

Which results in the check box for that log being pre-checked, and the
text '#53456' being added, which turns into a pretty link on GitHub.

It might be nice to group issues as well, but I didn't want to mess with
the chronological ordering.

For golang/go#52653.

Change-Id: If4615cd798ba72c1c1ee3cb43f1d1ad6d4319528
Reviewed-on: https://go-review.googlesource.com/c/build/+/425075
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Heschi Kreinick <heschi@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>
Reviewed-by: Bryan Mills <bcmills@google.com>