internal/relui/protos: update wire protocol
Development and testing of the release signing workflow helped us find
ways to adjust the wire protocol to be a better fit for our needs.

The responsibility to come up with a job ID for future requests is moved
to the initial sign request creation transaction, replacing the need for
taskName and retryCount parameters.

Have just one build type for macOS signing because we no longer need to
keep the AMD64 and ARM64 versions separate.

Add a message and endpoint for marking a previously started signing job
as obsolete, so it'll be possible to do this automatically when a relui
release workflow happens to be stopped by a release coordinator.

Update the sign server and its tests accordingly.

For golang/go#53632.

Change-Id: I99081b0a0f8140b68e1a4b20bc5e422516052e28
Reviewed-on: https://go-review.googlesource.com/c/build/+/428057
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>