internal/relui: enable retries for source match
While working on this week's security release, it occurred to me that
there was no graceful way to recover from a mistake cherrypicking fixes.
Because we picked the commit to look at in a separate task, that commit
would be locked in by the time the check failed, and retrying it
wouldn't accomplish anything.

Instead, read the branch head in the same task. Now we can fix mistakes
however we need to (force push, more cherrypicks) and just retry as
usual.

This does mess with the division of labor between build and version
task structs. So it goes.

Change-Id: Ia2022e732e71b32b7c2cb5e892083bcf8f5c6444
Reviewed-on: https://go-review.googlesource.com/c/build/+/429535
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>