internal/coordinator/pool/queue: prioritize release branches
There's some kind of bug with the coordinator that's causing it to miss
release branch work. Once it eventually shows up, it starves because
it's old.

As a workaround, prioritize release branches over others. This may not
be a good idea long-term but I'm not sure. The policy we really want is
probably too complicated to express as a total ordering and I don't know
if we want to take on that complexity.

For golang/go#48523.

Change-Id: I04a88ebaeb90636e478460b4cf71021a264151a6
Reviewed-on: https://go-review.googlesource.com/c/build/+/443342
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Heschi Kreinick <heschi@google.com>
Reviewed-by: Jenny Rakoczy <jenny@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>