internal/releasetargets: remove Race config from 1.20
Go 1.20 will not be shipping with precompiled .a files for most packages
in the distribution in favor of having them compiled as needed and
cached. The .a files for libraries that depend on cgo are left in so
that packages that depend on them can still be built without a C
compiler. With this cl, we will further stop producing .a files for race
targets (including those that use cgo). Because the race targets aren't
used for production releases, this will allow us to test if we run into
issues removing all precompiled .a files in a future release while
limiting impact, and will reduce the size of the distribution still
further.

For golang/go#47257

Change-Id: I686a74e9438785cb85d4ea9a8cc7ca6ca7c9c43a
Reviewed-on: https://go-review.googlesource.com/c/build/+/446735
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Michael Matloob <matloob@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Michael Matloob <matloob@golang.org>