env/linux-x86-sid: add cloudbuild configuration
Also, update to the latest tag rather than one from 2019. This is in
preparation for changing the host config to use a specific image tag
instead of baking the image tag into every docker image.

This change will allow us to keep a separate 'cloudbuild' tag from the
tag used for the main bulder. This tag can be updated weekly, allowing
us to run a second builder in parallel to determine when we want to
update to a newer image.

For golang/go#56673

Change-Id: I906a95b447949deba0b108b3009c440aa5c31171
Reviewed-on: https://go-review.googlesource.com/c/build/+/449037
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Jenny Rakoczy <jenny@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>