perf: improve regressions page on performance dashboard
This change makes a few improvements to the regressions page:
1. It binds regressions to benchmark+unit, so now the regressions page
   might break up the groupings compared to before (which is probably
   OK).
2. This binding lets us avoid putting regression information into the
   benchmark name, which previously broke the title's link, so this
   change also moves that information to just above each graph.
3. This change adds a link to the regression information which jumps to
   the offending commit on the per-benchmark+unit page.
4. It adds a file-an-issue link to the regressions page which
   streamlines some of the overhead of triaging.

Note that deltaScore isn't exported to the regressions page anymore
because JSON can't tolerate negative infinities, and the new code is
trying to be as simple as possible. The next step is probably to filter
out non-regressions entirely, rather than try to make the deltaScore
work. We can also always bring it back as a string if we feel it's very
useful.

Change-Id: Ib7cf16ef92534fd1d9b9c23798e9474e20f4e8c2
Reviewed-on: https://go-review.googlesource.com/c/build/+/459518
Run-TryBot: Michael Knyszek <mknyszek@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Michael Pratt <mpratt@google.com>