dashboard: update js-wasm-node18 known issue and limit it to Go 1.21+
The js/wasm port is compatible with Node.js 14 but not Node.js 18, and
we're in the middle of a freeze so it's too late to make changes now.
Target the new builder for Go 1.21 and newer, and replace the "add
builder" known issue with umbrella tracking issue for Node 18.

Keep the current js-wasm builer with Node 14 as is so that the port
doesn't deteriorate further in the meantime. It can be disabled for tip
as soon as we have CLs to make all.bash pass with Node 18 all prepared.

Fixes golang/go#57017.
Updates golang/go#57614.

Change-Id: I66cb4210f1e899329550162242c72de0da7aedda
Reviewed-on: https://go-review.googlesource.com/c/build/+/460575
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>