internal/task: read 6 weeks of dashboard history
Fix a hole in greenness checking. If:
- the repository has been modified since the last tag
- but not in the last week
- and none of its dependencies have changed
Then the commit we want to tag will be more than a week old and we won't
see it, which triggers the race-prevention logic that wants to see the
commit on the dashboard before proceeding.

Expand the window to 6 weeks, which will cover all the time since the
previous tag run and then some.

Change-Id: Ie5062d98f70a02ad77b85ca26ecae58cf969046e
Reviewed-on: https://go-review.googlesource.com/c/build/+/466437
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Heschi Kreinick <heschi@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>