cmd/relui: add a "ping early-in-cycle issues" workflow
When adding a dedicated relui workflow, it was tempting to make it fancy
with each issue being a sub-task, etc. But we wouldn't want that if/when
this task joins others as part of a "reopen the dev tree" workflow, so
keep it a simple atomic (and safe to re-run) task. This is similar to
how pushing issues or checking for release blockers is one atomic task.

For golang/go#58856.

Change-Id: Idc8ef3cd62b7bb98c97b37d38ef9cdc2beccc24c
Reviewed-on: https://go-review.googlesource.com/c/build/+/473160
Reviewed-by: Heschi Kreinick <heschi@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>