main.star: reorganize builder definitions and change builder names
This change reorganizes builder definitions to more closely resemble
what the finished product will look like: generating builders from data.
The relevant data to generate a builder is:
- LUCI bucket
- go.googlesource.com project
- go.googlesource.com/go branch to run against
- Builder type, which consists of:
  - Host (GOOS, GOARCH, + extra)
  - A series of modifications to the build and/or test.

It changes builder names to have the following format, which makes use
of the above information.

    $BUCKET/(x_$PROJECT-)?$GO_BRANCH-$BUILDER_TYPE

$GO_BRANCH always has the form "go<branch short name>". For example:
"gotip" or "go1.20".

(x_$PROJECT-) is omitted for the main Go repository because what we're
testing is evident from $GO_BRANCH (and the "x_" prefix is precisely to
indicate that a golang.org/x/* repository).

Where $BUILDER_TYPE expands to $HOST(-$RUN_MOD)* and $HOST expands to
$GOOS-$GOARCH(-$HOST_SPECIFIER)?.

A host maps to Swarming dimensions describing the type of machine used
for the build.

From all this information, the full list of builders is generated by
taking the product of all projects, Go branches, and builder types, then
filtering out builders that won't run in presubmit (no builder
definition in the "try" bucket) or that won't run in postsubmit (no
builder definition in the "ci" bucket). The filter function for
presubmit and postsubmit is present, but currently says "yes" for all
builders. In the future we expect this to change.

Change-Id: If3d58f7838ccd9d60d791036a4ec7a68dc8d7a3a
Reviewed-on: https://go-review.googlesource.com/c/build/+/474775
TryBot-Bypass: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Heschi Kreinick <heschi@google.com>