cmd/releaseschedule: add dark mode aware svg css
The SVG image used on https://go.dev/s/release is mostly unreadable if
viewed in dark mode (set in OS, browser, or github settings) as the black
text/lines on dark background are too close in color. It would be possible to
set a background color on the svg, but I think the better answer is CSS in
the SVG which is @media aware of prefers-color-scheme:dark to switch the
colors from black to white.

Remove the explicit style on line/path and add css for text/line/path which
is black but in dark mode changes to white.

Updates golang/go#58820
Fixes golang/go#59275