cmd/releaseschedule: basic doc and README.md
"cmd/releaseschedule" had no docs and no direction on what to do with
the output. Would be better to have links to the wiki page and basic
information on what to do with the output. Helps with discoverability as
well if someone searches for information that SVG.

Add package doc with request to update wiki & links to wiki. Generate
README.md with update-readmes.go.

Fixes golang/go#59314