all: regenerate READMEs with update-readmes.go
Based on the existance of update-readmes.go and the context in the commit
messages making changes to that file I expect that it should be run
periodically. Generation of new README.md files does not appear to have been
done for well over a year.

Run update-readmes.go to generate missing READMEs.