internal/relui: create module files for releases
Building on relui's new support for distpack builds, publish the
toolchain module files it generates. This turns out to be quite a
bit more complicated than I had hoped.

First, save the entire distpack archive, not just the binary
distribution. Extract the binary distribution for use in the normal
workflow. Then extract the modules files (really just the .zip, the
.mod and .info are tiny) to a separate artifact. Publish those to the
CDN alongside the normal release artifacts.

For non-distpack builds, copy the dl2mod code in from CL 475920 and
generate the module files as a postprocessing step on the binary
distribution.

The really annoying part is macOS, where we need to distribute the
signed binaries in the module, not the ones we built. For that, take the
module zip generated by distpack and replace the binaries with the ones
from the signed binary distribution. This is a fairly picky process, but
hopefully means we don't risk drifting out of sync with the canonical
distpacks.

Finally, the testing changes: the fake make.bash needs to emulate
distpack too, which wasn't too bad. And then the new requirement to
process the results of the macOS signing process means that the testing
fakes can't just produce cute text files -- they have to contain the
real distribution. Change the fake macOS packaging tools to deal in
prefixed tarballs.

For golang/go#58659.

Change-Id: I9cca8af2cb562fb39ef805a044109877b4ab7fb8
Reviewed-on: https://go-review.googlesource.com/c/build/+/482996
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>