cmd/relui: address performance issues with module zips
Creating the module zips is surprisingly CPU-intensive, and we end up
doing it many times in parallel, resulting in it taking more than 10
minutes to finish and triggering the watchdog timer.

Disable the watchdog for this task, and give relui more CPU so that it's
less of a problem in the first place. (Collateral damage to other tasks
is still a concern.)

For golang/go#58659.

Change-Id: I65b5c47f9fdc011885c0d5f886f4050551d55f61
Reviewed-on: https://go-review.googlesource.com/c/build/+/489275
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Run-TryBot: Heschi Kreinick <heschi@google.com>