perf: add PGO comparisons to dashboard
CL 498278 starts collecting PGO benchmark results with key pgo:on. This
CL adds PGO results to the dashboard via two different comparison
dimensions:

1. "/pgo=on,toolchain:baseline-vs-experiment" is "toolchain:baseline" vs
   "toolchain:experiment" when "pgo:on". i.e., how much has performance
   changed since baseline if PGO has been enabled the entire time.

2. "/toolchain:experiment,pgo=off-vs-on" is "pgo:off" vs "pgo:on" for
   "toolchain:experiment". i.e., how much performance changes when
   enabling PGO at experiment.

These names are added as a suffix to the existing benchmark name. The
unmodified benchmark name remains the same as before:
"toolchain:baseline" vs "toolchain:experiment" when "pgo:off".

I would argue that tracking (1) makes the most sense for a few reasons:

a. It is easier to understand the charts. If we make a non-PGO
optimization, both the standard (non-PGO) and (1) charts will show
improvement. If we make a PGO optimization, the standard chart will be
unchanged and (1) will show improvement. If we track (2), a non-PGO
optimization may show a regression (percentage closer to zero but still
negative) because the slightly faster experiment build has a smaller
percentage improvement when PGO is applied. This looks like things are
worse even though the overall binary is still faster.

b. The ultimate goal of PGO is to save money by making more efficient
binaries, which is what (1) shows, not to make a big delta vs PGO
disabled. The delta from PGO disabled is effectively "marketing" to show
folks it is worthwhile to switch it on.

That said, both can be interesting views, so this CLs enables both while
we gain experience with them.

Adding PGO runs is expected to increase benchmarking time to closer to
8hr. Bump the timeout to be safe.

Change-Id: I974aa84c03ac8df4b4c483c57a98805edc708574
Reviewed-on: https://go-review.googlesource.com/c/build/+/498335
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Run-TryBot: Michael Pratt <mpratt@google.com>