cmd/gopherbot: close WaitingForInfo issues as "not planned"
GitHub now makes it possible to specify a reason when closing an issue,
which can make it easier to tell at a glance that something was closed
as duplicate, stale, because it doesn't reproduce, etc., as opposed to
because the reported problem was resolved.

Most issues that time out in WaitingForInfo are closed because there's
insufficient information and the author is unresponsive, so close them
as "not planned" by default. It's always possible to re-close an issue
as resolved when applicable, but that's the exceptional case.

Also use a newer Go toolchain to build while here.

Fixes golang/go#60761.

Change-Id: I5572d76b2b1d240de90ce5a602c1965a739ff84d
Reviewed-on: https://go-review.googlesource.com/c/build/+/502735
Reviewed-by: Carlos Amedee <carlos@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>