cmd/relui: add Cloud Build triggering support
Build the Google-facing BoringCrypto image as part of publishing the
release.

I found myself a bit stuck with modeling the step. Currently it's split
in two, which means it's unlikely to trigger two builds but we don't
have any way to retry a build if it flakes. Merging the creation and the
await lets us retry, but creates a risk of publishing the Docker image
twice. Splitting the publishing step out would fix that, but I didn't
want to have to write more code to do the publishing. I think we'll have
to see how it goes.

Change-Id: I9ad4e79d05611ea744191a4d9d4a6c8df38c523a
Reviewed-on: https://go-review.googlesource.com/c/build/+/504819
Auto-Submit: Heschi Kreinick <heschi@google.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>