internal/task: replace use of go.dev/dl/?mode=json API with local params
The tweet tasks are some of the earliest we've automated for relui.
In fact, they were ready for use with the release process ahead of
the rest of relui, so the only place to add them was the CLI-based
x/build/cmd/releasebot command.

Since release coordinators needed to provide all of the task inputs as
a JSON object over CLI, at the time, it was easier to fetch information
about the Go release artifacts that had been published for the very Go
release being made by making a network call to the go.dev/dl/?mode=json
API.

By now, relui is complete and has all the steps interconnected without
toil of invoking individual release steps manually, so it's easy for it
to pass the release artifacts to the tweeting tasks directly, and avoid
having to ask the website for the release it itself published (silly!).

We can also drop the lightweight checking of bad version strings done
in oneOrTwoGoVersions. Those typos can't happen now since relui picks
the next version to release automatically.

This CL does the re-wiring to drop the use of API from TweetRelease but
makes almost no changes to the current comprehensive tweet tests, and
they're still passing. The next change simplifies the tests to drop
their temporary use of the API.

For golang/go#57062.

Change-Id: I60efc8c8c96152ff52c9fb5cc0f91262f0d8f494
Reviewed-on: https://go-review.googlesource.com/c/build/+/504519
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>