internal/releasetargets: reapply "set LongTestBuilder for darwin-amd64"
This reapplies CL 503758, which was rolled back in CL 503516.

Trying again after scaling watchdog timeout to take into account
that this is a longtest builder and it's expected to take longer, and
after doubling its timeout scale value in the builder configuration.
Since it's faster than signing, it's not expected to have any effect
on the speed of the release process.

We also have the option of using the "skip testing" workflow
parameter if needing to temporarily skip it again, something
I forgot about when rolling it back earlier.

For golang/go#29252.
For golang/go#60919.

Change-Id: I179479b46a31ec5e95d4e9cd8b1f8fedb765880c
Reviewed-on: https://go-review.googlesource.com/c/build/+/504524
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>