env,cmd/buildlet/stage0: add bootstrapswarm to stage0 layer
We want to move our Docker-based builders over to LUCI. Rather than
building a second set of images, add the LUCI bootstrap command to the
stage0 layer, along with a script that selects which to run based on the
GCE project ID.

It's sort of silly to put this in cmd/buildlet, but I didn't want to
make another image and then have to teach xb about it, or undertake a
big yak shave to get rid of xb. Since the image is built from the root
of the module it doesn't matter much.

This doesn't do any of the necessary steps to make LUCI work once the
bootstrap command runs. That'll come later.

Change-Id: Iaa911cf86a9f52efd688836cfaa6f0e4a11ed884
Reviewed-on: https://go-review.googlesource.com/c/build/+/506876
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Auto-Submit: Heschi Kreinick <heschi@google.com>