internal/{relui,task,workflow}: check coordinator parameter sooner
It's not often that a bad parameter value is provided to a workflow.
But when it does happen, it'd be nicer if relui would tell you about
the problem right away. Instead, it quietly starts a workflow that
will predictably fail at some arbitrary later time, only after it
reaches the first task that uses said parameter.

This is especially true for cases that are trivial for relui to check
automatically, such as a typo in release coordinator username or name
or Gerrit email not being in gophers package. So add a check to report
right away when a release coordinator username parameter will result in
an error. This lets humans spend their attention on more interesting
things (in the long run).

Change-Id: If12b6275d56b6b03adc27958c1792391ba600712
Reviewed-on: https://go-review.googlesource.com/c/build/+/508435
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Cherry Mui <cherryyz@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>