main.star: fix project name passed into golangbuild for go-internal repo
The name of the invoking project is passed into the golangbuild luciexe
through properties. golangbuild relies on it in many testing logics.

Previous implementation was not aware of above implication and
misassumed that the name was only used for display and can be changed
freely. This CL fixes it.

Change-Id: Ic077ba451a338d3f5a265ddcdef54b17a211b67e
Reviewed-on: https://go-review.googlesource.com/c/build/+/512035
Reviewed-by: Heschi Kreinick <heschi@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>
TryBot-Bypass: Heschi Kreinick <heschi@google.com>