main.star: set test_shard to a valid value by default
Currently test_shard is set to an invalid test shard (0 shards, 0 shard
ID) by default. But golangbuild assumes that if there's no sharding, it
can just skip passing this property if sharding isn't needed. That may
need to change, but this is a quick fix that will resolve issues with
the darwin builders.

Change-Id: If174aaac261858b645be67b01b2164b45a92d8d1
Reviewed-on: https://go-review.googlesource.com/c/build/+/513075
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Bypass: Heschi Kreinick <heschi@google.com>