cmd/gorebuild: add tool to reproduce posted Go binaries
This command rebuilds or verifies all the artifacts posted on
go.dev/dl for the latest supported releases (the last patch of
the last two major releases, plus the most recent release candidate
if we're approaching a new release).

It is meant to be run by the Go team to update a status page
that can be linked from reproducible-builds.org, but it is also
meant to be run by anyone who wants to "trust but verify" the
status page themselves.

For golang/go#57120.
For golang/go#58884.

Change-Id: I80a70275c1821a66b6219d24f29c2d11bfe464a8
Reviewed-on: https://go-review.googlesource.com/c/build/+/513975
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Run-TryBot: Russ Cox <rsc@golang.org>