main.star: make the no-network check required, turn on in main Go repo
The initial version of the no-network check failed open: if golangbuild
didn't see unshare (or ip) in $PATH, it considered the host as not
capable of implementing the no-network check and wouldn't turn it on.

Even though we can see if the check works by looking at a particular
build log, it's better to remove the silent possibility of 0 builders
where it actually turns on. Do that explicitly on linux-386 and -amd64
builder types.

By now all known tests that needed attention in the main Go repository
have been taken care of. The current implementation of the no-network
check has more room for improvement, but it's still better than the
one in x/build. Let it serve us until we have root builders in LUCI
and decide to invest into further improvements for this check.

Change-Id: I6c50275fc68799db9adb21cf9036ebba623d8bc1
Reviewed-on: https://go-review.googlesource.com/c/build/+/513399
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
TryBot-Bypass: Dmitri Shuralyov <dmitshur@golang.org>