main.star: categorize projects and select builders
Categorize each project as a particular type, and use that type to
select which builds should run. Unwanted builds are removed from
postsubmit and made opt-in in presubmit. (It might be reasonable to
remove them from presubmit too, but there's not too much harm in leaving
them in, and you never know what'll end up being useful?)

CORE and LIBRARY are currently treated the same, but I think it's a real
distinction that will become clearer as we continue adding builders. For
example, we might want to not spend one-off machines' time testing them.

The enums are janky. I found
https://groups.google.com/g/bazel-dev/c/o_giPbwcPc0 but didn't
understand it, which seemed like a reason not to get fancy.

Change-Id: Ia444b01f1f4e8f38948e0eab29a263742d30bb58
Reviewed-on: https://go-review.googlesource.com/c/build/+/515355
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Bypass: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Auto-Submit: Heschi Kreinick <heschi@google.com>